refactor: move crd clients to crd_client folder
